fix: modify export for the package
