+++ types.d.ts
@@ -1,6 +1,3 @@
-/**
- * @module @adobe/aio-lib-core-networking
- */
 declare module "@adobe/aio-lib-core-networking" {
     /**
      * This class provides methods to implement fetch with retries.
@@ -11,13 +8,12 @@ declare module "@adobe/aio-lib-core-networking" {
         /**
          * This function will retry connecting to a url end-point, with
          * exponential backoff. Returns a Promise.
-         *
-         * @param {string} url endpoint url
-         * @param {object} requestOptions request options
-         * @param {object} retryOptions retry options with keys being maxRetries and initialDelay in ms
-         * @param {Function|Array} [retryOn] Optional Function or Array. If provided, will be used instead of the default
-         * @param {Function|number} [retryDelay] Optional Function or number. If provided, will be used instead of the default
-         * @returns {Promise<Response>} Promise object representing the http response
+         * @param url - endpoint url
+         * @param requestOptions - request options
+         * @param retryOptions - retry options with keys being maxRetries and initialDelay in ms
+         * @param [retryOn] - Optional Function or Array. If provided, will be used instead of the default
+         * @param [retryDelay] - Optional Function or number. If provided, will be used instead of the default
+         * @returns Promise object representing the http response
          */
         exponentialBackoff(url: string, requestOptions: any, retryOptions: any, retryOn?: ((...params: any[]) => any) | any[], retryDelay?: ((...params: any[]) => any) | number): Promise<Response>;
     }
